+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - set-attribute: to set pot attributes (options/flags/configurations)
 - get-attribute: to get pot attributes (options/flags/configurations)
 - FreeBSD version usable to create a pot are all the ones listed in the FreeBSD MANIFEST
+- attributes: add persistent attribute to jail
 
 ### Changed
 - inherit network: added ipv6 support (automatic)